archived-architecture/ai-serving-api/requirements.txt
langchain
google-cloud-aiplatform
pymongo